refactor: SpinnerAdapter 클래스 삭제 Spinner를 acivity에서 생성하도록 수정
